feat: 为所有算子确定构造器以免少传参数

+++ src/03frontend/include/frontend/operator.h
@@ -12,8 +12,8 @@ namespace refactor::frontend {
 
     using Int = long long;
     using Ints = std::vector<long long>;
-    using Float = double;
-    using Floats = std::vector<double>;
+    using Float = float;
+    using Floats = std::vector<float>;
     using String = std::string;
     using Strings = std::vector<std::string>;
     using Tensor_ = std::shared_ptr<Tensor>;
